requirements.txt
Flask==3.1.0
dropbox~=12.0.2
python-dotenv~=1.0.1
Pillow
pytesseract~=0.3.13
pdf2image~=1.17.0
PyPDF2~=3.0.1
requests~=2.32.3
schedule~=1.2.2
openai==1.54.4
slack_sdk
celery~=5.4.0
PyAutoGUI~=0.9.54
psutil~=6.1.0
PyWinCtl~=0.4.1
